move view_helper to lib/
