+++ app/lib/primer/view_helper.rb
@@ -1,21 +1,3 @@
 # frozen_string_literal: true
 
-module Primer
-  # Module to allow shorthand calls for Primer components
-  module ViewHelper
-    class ViewHelperNotFound < StandardError; end
-
-    HELPERS = {
-      octicon: "Primer::OcticonComponent",
-      heading: "Primer::HeadingComponent",
-      time_ago: "Primer::TimeAgoComponent",
-      image: "Primer::Image"
-    }.freeze
-
-    HELPERS.each do |name, component|
-      define_method "primer_#{name}" do |*args, **kwargs, &block|
-        render component.constantize.new(*args, **kwargs), &block
-      end
-    end
-  end
-end
+require "primer/view_helper"
