+++ lib/primer/view_components.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "primer/classify"
+require "primer/view_helper"
 require "primer/view_components/version"
 require "primer/view_components/engine"
